refactor: rename and relocate receivePersonaData function for better organization
