+++ packages/persona/index.ts
@@ -1,2 +1,2 @@
 export { getPersonasByPersonaId } from "./src/getPersonasByPersonaId.js";
-export { receivePersonaData } from "./src/internal.js";
+export { receivePersonaData } from "./src/receivePersonaData.js";
