+++ packages/persona/src/handleSelectGamePersona.ts
@@ -13,7 +13,6 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
 import { NPSMessage, ServerLogger } from "rusty-motors-shared";
 
 /**
@@ -22,7 +21,11 @@ import { NPSMessage, ServerLogger } from "rusty-motors-shared";
  * @param {ServerLogger} log
  * @returns {Promise<NPSMessage>}
  */
-export async function handleSelectGamePersona(requestPacket: NPSMessage, log: ServerLogger): Promise<NPSMessage> {
+
+export async function handleSelectGamePersona(
+	requestPacket: NPSMessage,
+	log: ServerLogger,
+): Promise<NPSMessage> {
 	log.debug("_npsSelectGamePersona...");
 
 	log.debug(
